requirements.txt
adbutils==1.0.1
apkutils2==1.0.0
asgiref==3.3.4
blinker==1.4
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cigam==0.0.3
click==7.1.2
cryptography==3.3.2
decorator==5.1.1
deprecation==2.1.0
Flask==1.1.4
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
kaitaistruct==0.9
ldap3==2.8.1
MarkupSafe==2.0.1
mitmproxy==6.0.0
msgpack==1.0.4
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
pip==21.2.3
protobuf==3.14.0
publicsuffix2==2.20191221
pure-python-adb==0.3.0.dev0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pyelftools==0.28
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
requests==2.28.1
retry==0.9.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
setuptools==57.4.0
six==1.16.0
sortedcontainers==2.3.0
tornado==6.2
urllib3==1.26.10
urwid==2.1.2
Werkzeug==1.0.1
whichcraft==0.6.1
wsproto==1.0.0
xmltodict==0.13.0
zipp==3.8.1
zstandard==0.14.1
